refactor(docs): move site.config.tsx values to lib/constants.ts
Plain data (github, nav, prompt) doesn't need JSX or a root-level
config file. Moved to lib/constants.ts and deleted site.config.tsx.
