+++ apps/docs/components/geistdocs/github-button.tsx
@@ -1,5 +1,5 @@
 import { SiGithub } from "@icons-pack/react-simple-icons";
-import { github } from "@/site.config";
+import { github } from "@/lib/constants";
 import { Button } from "../ui/button";
 
 export const GitHubButton = () => {
